refactor:  Userのディレクトリ整理 (#615)
* Userのディレクトリ整理(wip)

* ah

* mvmv

* fx
