+++ src/User/findUser.resolver.ts
@@ -1,14 +1,14 @@
 import { GraphQLError } from "graphql";
 
-import { QueryResolvers } from "../../graphql.js";
-import { ResolverDeps } from "../../types.js";
+import { QueryResolvers } from "../resolvers/graphql.js";
+import { ResolverDeps } from "../resolvers/types.js";
 
-export const findUser = ({ logger, userRepository }: Pick<ResolverDeps, "logger" | "userRepository">) =>
+export const resolverFindUser = ({ logger, userService }: Pick<ResolverDeps, "logger" | "userService">) =>
   (async (_parent, { input: { name } }, { currentUser: ctxUser }, info) => {
     if (!name) {
       logger.error({ path: info.path, args: { input: { name } }, userId: ctxUser?.id }, "Invalid input");
       throw new GraphQLError("name must be provided"); // TODO: error messsage
     }
 
-    return userRepository.findByName(name);
+    return userService.findByName(name);
   }) satisfies QueryResolvers["findUser"];
