+++ src/resolvers/Mutation/changeUserDisplayName/resolver.ts
@@ -1,9 +1,9 @@
 import { MutationResolvers, ResolversTypes } from "../../graphql.js";
 import { ResolverDeps } from "../../types.js";
 
-export const resolverChangeUserDisplayName = ({ userRepository }: Pick<ResolverDeps, "userRepository">) =>
+export const resolverChangeUserDisplayName = ({ userService }: Pick<ResolverDeps, "userService">) =>
   (async (_parent, { renameTo }, { currentUser }) => {
-    const user = await userRepository.changeDisplayName(currentUser.id, renameTo);
+    const user = await userService.changeDisplayName(currentUser.id, renameTo);
     return {
       __typename: "ChangeUserDisplayNameSucceededPayload",
       user,
