--- src/resolvers/Query/findUser/findUser.graphql
@@ -1,7 +1,0 @@
-type Query {
-  findUser(input: FindUserInput!): User
-}
-
-input FindUserInput {
-  name: String
-}
