--- src/resolvers/Query/getUser/getUser.graphql
@@ -1,3 +1,0 @@
-type Query {
-  getUser(id: ID!): User!
-}
