+++ src/resolvers/Query/whoami/resolver.ts
@@ -3,8 +3,8 @@ import { GraphQLError } from "graphql";
 import { QueryResolvers } from "../../graphql.js";
 import { ResolverDeps } from "../../types.js";
 
-export const resolverWhoami = ({ userRepository }: Pick<ResolverDeps, "userRepository">) =>
+export const resolverWhoami = ({ userService }: Pick<ResolverDeps, "userService">) =>
   (async (_parent, _args, { currentUser }) =>
-    userRepository.getById(currentUser.id).catch(() => {
+    userService.getById(currentUser.id).catch(() => {
       throw new GraphQLError("Internal server error");
     })) satisfies QueryResolvers["whoami"];
