--- src/resolvers/User/User.graphql
@@ -1,11 +1,0 @@
-type User implements Node {
-  id: ID!
-  name: String!
-  displayName: String!
-  icon: String!
-
-  mylist(id: ID!): Mylist
-
-  isEditor: Boolean! @deprecated(reason: "Use hasRole instead.")
-  isAdministrator: Boolean! @deprecated(reason: "Use hasRole instead.")
-}
