--- src/resolvers/User/hasRole/hasRole.graphql
@@ -1,8 +1,0 @@
-type User {
-  hasRole(role: UserRole!): Boolean
-}
-
-enum UserRole {
-  EDITOR
-  ADMIN
-}
