--- src/resolvers/User/likes/likes.graphql
@@ -1,3 +1,0 @@
-type User {
-  likes: Mylist
-}
