--- src/resolvers/User/mylists/mylists.graphql
@@ -1,17 +1,0 @@
-type User {
-  mylists(
-    first: Int
-    after: String
-
-    last: Int
-    before: String
-
-    range: [MylistShareRange!]! = [PUBLIC]
-
-    orderBy: UserMylistsOrderBy! = { createdAt: DESC }
-  ): MylistConnection!
-}
-
-input UserMylistsOrderBy {
-  createdAt: SortOrder
-}
