--- src/resolvers/User/nicovideoRegistrationRequests/nicovideoRegistrationRequests.graphql
@@ -1,15 +1,0 @@
-type User {
-  nicovideoRegistrationRequests(
-    first: Int
-    after: String
-
-    last: Int
-    before: String
-
-    orderBy: UserNicovideoRegistrationRequestsOrderBy! = { createdAt: DESC }
-  ): NicovideoRegistrationRequestConnection!
-}
-
-input UserNicovideoRegistrationRequestsOrderBy {
-  createdAt: SortOrder
-}
